pyproject.toml
[tool.poetry]
name = "project-nova"
version = "0.1.1"
description = "A personal assistant project"
authors = ["Ankush Roy <ankushroy7926@gmail.com> Arghya Sarkar <arghyasarkar.nolan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyttsx3 = "^2.98"
pyaudio = "^0.2.14"
speechrecognition = "^3.11.0"
botocore = "^1.35.47"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
ollama = "^0.3.3"
pocketsphinx = "^5.0.3"
python-vlc = "^3.0.21203"
youtube-dl = "^2021.12.17"
yt-dlp = "^2024.10.22"

[tool.poetry.scripts]
nova = "project_nova.app:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
